+++ src/components/HomePage.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export const HomePage: React.FC = () => {
+  return (
+    <div className="section">
+      <div className="container">
+        <h1 className="title">Home page</h1>
+      </div>
+    </div>
+  );
+};
