+++ src/components/NotFoundPage.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export const NotFoundPage: React.FC = () => {
+  return (
+    <div className="section">
+      <div className="container">
+        <h1 className="title">Page not found</h1>
+      </div>
+    </div>
+  );
+};
