+++ src/components/TabsPage.tsx
@@ -1,0 +1,45 @@
+import classNames from 'classnames';
+import React from 'react';
+import { Link, useParams } from 'react-router-dom';
+import { Tab } from '../types/Tab';
+
+const tabs: Tab[] = [
+  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
+  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
+  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
+];
+
+export const TabsPage: React.FC = () => {
+  const { tabId } = useParams<{ tabId?: string }>();
+  const activeTab = tabs.find(tab => tab.id === tabId);
+
+  return (
+    <div className="section">
+      <div className="container">
+        <h1 className="title">Tabs page</h1>
+
+        <div className="tabs is-boxed">
+          <ul>
+            {tabs.map(tab => {
+              const isActive = tab.id === tabId;
+
+              return (
+                <li
+                  key={tab.id}
+                  data-cy="Tab"
+                  className={classNames({ 'is-active': isActive })}
+                >
+                  <Link to={`/tabs/${tab.id}`}>{tab.title}</Link>
+                </li>
+              );
+            })}
+          </ul>
+        </div>
+
+        <div className="block" data-cy="TabContent">
+          {!tabId || !activeTab ? 'Please select a tab' : activeTab.content}
+        </div>
+      </div>
+    </div>
+  );
+};
